refactor: refactor storage code (#1090)
* refactor: rename `PortalStorage` to `HistoryStorage`

* refactor: create `trin-storage` workspace crate

* refactor: trin-history storage

* refactor: trin-beacon storage

* refactor: trin-state storage

* refactor: move storage tests

* fix: remove unused crates from portalnet

* fix: address PR comments
