+++ Cargo.toml
@@ -37,6 +37,7 @@ tracing-subscriber = "0.3.15"
 trin-beacon = { path = "trin-beacon" }
 trin-history = { path = "trin-history" }
 trin-state = { path = "trin-state" }
+trin-storage = { path = "trin-storage" }
 trin-utils = { path = "trin-utils" }
 trin-validation = { path = "trin-validation" }
 utp-rs = "0.1.0-alpha.8"
@@ -60,6 +61,7 @@ members = [
     "trin-beacon",
     "trin-history",
     "trin-state",
+    "trin-storage",
     "trin-utils",
     "trin-validation",
     "utp-testing",
