+++ portalnet/Cargo.toml
@@ -32,9 +32,6 @@ lru = "0.7.8"
 parking_lot = "0.11.2"
 rand = "0.8.4"
 rlp = "0.5.0"
-rusqlite = { version = "0.26.3", features = ["bundled"] }
-r2d2 = "0.8.9"
-r2d2_sqlite = "0.19.0"
 serde = { version = "1.0.150", features = ["derive"] }
 serde_json = "1.0.89"
 smallvec = "1.8.0"
@@ -45,6 +42,7 @@ thiserror = "1.0.29"
 tokio = { version = "1.14.0", features = ["full"] }
 tracing = "0.1.36"
 trin-metrics = { path = "../trin-metrics" }
+trin-storage = { path = "../trin-storage" }
 trin-utils = { path = "../trin-utils" }
 trin-validation = { path = "../trin-validation" }
 validator = { version = "0.13.0", features = ["derive"] }
