+++ trin-beacon/Cargo.toml
@@ -19,10 +19,16 @@ ethportal-api = {path = "../ethportal-api"}
 eyre = "0.6.8"
 parking_lot = "0.11.2"
 portalnet = { path = "../portalnet" }
+r2d2 = "0.8.9"
+r2d2_sqlite = "0.19.0"
+rusqlite = { version = "0.26.3", features = ["bundled"] }
 light-client = { path = "../light-client" }
 serde_json = "1.0.89"
+ssz_types = "0.5.4"
 tokio = {version = "1.14.0", features = ["full"]}
 tracing = "0.1.36"
+trin-metrics = { path = "../trin-metrics" }
+trin-storage = { path = "../trin-storage" }
 trin-validation = { path = "../trin-validation" }
 trin-utils = { path = "../trin-utils" }
 utp-rs = "0.1.0-alpha.8"
