+++ trin-beacon/src/network.rs
@@ -4,7 +4,7 @@ use parking_lot::RwLock as PLRwLock;
 use tokio::sync::RwLock;
 use utp_rs::socket::UtpSocket;
 
-use crate::{sync::BeaconSync, validation::BeaconValidator};
+use crate::{storage::BeaconStorage, sync::BeaconSync, validation::BeaconValidator};
 use ethportal_api::{
     types::{distance::XorMetric, enr::Enr, portal_wire::ProtocolId},
     BeaconContentKey,
@@ -13,8 +13,8 @@ use portalnet::{
     config::PortalnetConfig,
     discovery::{Discovery, UtpEnr},
     overlay::{OverlayConfig, OverlayProtocol},
-    storage::{BeaconStorage, PortalStorageConfig},
 };
+use trin_storage::PortalStorageConfig;
 use trin_validation::oracle::HeaderOracle;
 
 /// Beacon network layer on top of the overlay protocol. Encapsulates beacon network specific data
