+++ trin-history/Cargo.toml
@@ -19,19 +19,27 @@ ethereum_ssz = "0.5.3"
 ethportal-api = {path = "../ethportal-api"}
 parking_lot = "0.11.2"
 portalnet = { path = "../portalnet" }
+r2d2 = "0.8.9"
+r2d2_sqlite = "0.19.0"
 serde_json = "1.0.89"
 tokio = { version = "1.14.0", features = ["full"] }
 tracing = "0.1.36"
 tree_hash = "0.5.2"
+trin-metrics = { path = "../trin-metrics" }
+trin-storage = { path = "../trin-storage" }
 trin-utils = { path = "../trin-utils" }
 trin-validation = { path = "../trin-validation" }
 utp-rs = "0.1.0-alpha.8"
 
 [dev-dependencies]
 env_logger = "0.9.0"
+quickcheck = "1.0.3"
 httpmock = "0.6.6"
+rand = "0.8.4"
 rstest = "0.18.2"
+serial_test = "0.5.1"
 ssz_types = "0.5.4"
 test-log = { version = "0.2.11", features = ["trace"] }
+tokio-test = "0.4.2"
 tracing-subscriber = "0.3.15"
 ureq = { version = "2.5.0", features = ["json"] }
