+++ trin-history/src/jsonrpc.rs
@@ -14,11 +14,11 @@ use ethportal_api::{
     utils::bytes::hex_encode,
     ContentValue, HistoryContentKey, OverlayContentKey, RawContentKey,
 };
-use portalnet::storage::ContentStore;
 use serde_json::{json, Value};
 use ssz::Encode;
 use tokio::sync::{mpsc, Mutex, RwLock};
 use tracing::error;
+use trin_storage::ContentStore;
 
 use crate::network::HistoryNetwork;
