+++ trin-history/src/lib.rs
@@ -4,6 +4,7 @@
 pub mod events;
 mod jsonrpc;
 pub mod network;
+mod storage;
 pub mod utils;
 pub mod validation;
 
@@ -24,8 +25,8 @@ use portalnet::{
     config::PortalnetConfig,
     discovery::{Discovery, UtpEnr},
     events::{EventEnvelope, OverlayRequest},
-    storage::PortalStorageConfig,
 };
+use trin_storage::PortalStorageConfig;
 use trin_validation::oracle::HeaderOracle;
 
 type HistoryHandler = Option<HistoryRequestHandler>;
