+++ trin-metrics/src/storage.rs
@@ -178,3 +178,54 @@ impl StorageMetricsReporter {
         }
     }
 }
+
+#[cfg(test)]
+#[allow(clippy::unwrap_used)]
+pub mod test {
+    use super::*;
+
+    #[test]
+    fn test_precision_for_percentage() {
+        fn formatted_percent(ratio: f64) -> String {
+            let precision = StorageMetricsReporter::precision_for_percentage(ratio * 100.0);
+            format!("{:.*}%", precision, ratio * 100.0)
+        }
+        assert_eq!(formatted_percent(1.0), "100%");
+        assert_eq!(formatted_percent(0.9999), "100%");
+        assert_eq!(formatted_percent(0.9949), "99%");
+
+        assert_eq!(formatted_percent(0.10001), "10%");
+        assert_eq!(formatted_percent(0.1), "10%");
+        assert_eq!(formatted_percent(0.09949), "9.9%");
+
+        assert_eq!(formatted_percent(0.010001), "1.0%");
+        assert_eq!(formatted_percent(0.01), "1.0%");
+        assert_eq!(formatted_percent(0.009949), "0.99%");
+
+        assert_eq!(formatted_percent(0.0010001), "0.10%");
+        assert_eq!(formatted_percent(0.001), "0.10%");
+        assert_eq!(formatted_percent(0.0009949), "0.099%");
+
+        assert_eq!(formatted_percent(0.00010001), "0.010%");
+        assert_eq!(formatted_percent(0.0001), "0.010%");
+        assert_eq!(formatted_percent(0.00009949), "0.0099%");
+
+        assert_eq!(formatted_percent(0.000010001), "0.0010%");
+        assert_eq!(formatted_percent(0.00001), "0.0010%");
+        assert_eq!(formatted_percent(0.0000095), "0.0010%");
+        assert_eq!(formatted_percent(0.00000949), "0.0009%");
+
+        assert_eq!(formatted_percent(0.0000010001), "0.0001%");
+        assert_eq!(formatted_percent(0.000001), "0.0001%");
+        assert_eq!(formatted_percent(0.0000009949), "0.0001%");
+        assert_eq!(formatted_percent(0.0000005001), "0.0001%");
+        assert_eq!(formatted_percent(0.0000004999), "0.0000%");
+        assert_eq!(formatted_percent(0.0), "0.0000%");
+
+        // We mostly care that values outside of [0.0, 1.0] do not crash, but
+        // for now we also check that they pin to 0 or 4.
+        assert_eq!(StorageMetricsReporter::precision_for_percentage(101.0), 0);
+        assert_eq!(StorageMetricsReporter::precision_for_percentage(-0.001), 4);
+        assert_eq!(StorageMetricsReporter::precision_for_percentage(-1000.0), 4);
+    }
+}
