+++ trin-state/Cargo.toml
@@ -18,8 +18,12 @@ ethereum-types = "0.14.1"
 ethportal-api = { path = "../ethportal-api" }
 parking_lot = "0.11.2"
 portalnet = { path = "../portalnet" }
+r2d2 = "0.8.9"
+r2d2_sqlite = "0.19.0"
 tracing = "0.1.36"
 tokio = {version = "1.14.0", features = ["full"]}
+trin-storage = { path = "../trin-storage" }
+trin-metrics = { path = "../trin-metrics" }
 trin-validation = { path = "../trin-validation" }
 utp-rs = "0.1.0-alpha.8"
