+++ trin-state/src/network.rs
@@ -3,6 +3,7 @@ use std::sync::Arc;
 use tokio::sync::RwLock;
 use utp_rs::socket::UtpSocket;
 
+use crate::storage::StateStorage;
 use ethportal_api::{
     types::{distance::XorMetric, enr::Enr, portal_wire::ProtocolId},
     StateContentKey,
@@ -11,17 +12,17 @@ use portalnet::{
     config::PortalnetConfig,
     discovery::{Discovery, UtpEnr},
     overlay::{OverlayConfig, OverlayProtocol},
-    storage::{PortalStorage, PortalStorageConfig},
 };
 use trin_validation::oracle::HeaderOracle;
 
 use crate::validation::StateValidator;
+use trin_storage::PortalStorageConfig;
 
 /// State network layer on top of the overlay protocol. Encapsulates state network specific data and
 /// logic.
 #[derive(Clone)]
 pub struct StateNetwork {
-    pub overlay: Arc<OverlayProtocol<StateContentKey, XorMetric, StateValidator, PortalStorage>>,
+    pub overlay: Arc<OverlayProtocol<StateContentKey, XorMetric, StateValidator, StateStorage>>,
 }
 
 impl StateNetwork {
@@ -32,7 +33,7 @@ impl StateNetwork {
         portal_config: PortalnetConfig,
         header_oracle: Arc<RwLock<HeaderOracle>>,
     ) -> anyhow::Result<Self> {
-        let storage = Arc::new(PLRwLock::new(PortalStorage::new(
+        let storage = Arc::new(PLRwLock::new(StateStorage::new(
             storage_config,
             ProtocolId::State,
         )?));
