+++ trin-storage/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "trin-storage"
+version = "0.1.0"
+edition = "2021"
+repository = "https://github.com/ethereum/trin/tree/master/trin-storage"
+license = "GPL-3.0"
+readme = "README.md"
+keywords = ["ethereum", "portal-network"]
+categories = ["cryptography::cryptocurrencies"]
+description = "Storage library for Trin."
+authors = ["https://github.com/ethereum/trin/graphs/contributors"]
+
+[dependencies]
+anyhow = "1.0.68"
+discv5 = { version = "0.4.0", features = ["serde"] }
+ethportal-api = {path = "../ethportal-api"}
+rusqlite = { version = "0.26.3", features = ["bundled"] }
+r2d2 = "0.8.9"
+r2d2_sqlite = "0.19.0"
+tracing = "0.1.36"
+thiserror = "1.0.29"
